requirements.txt
Pillow
pytesseract
openai
pandas
python-dotenv